requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pillow==10.1.0
six==1.16.0
numpy==1.26.2
python-multipart==0.0.6